Replace slot-fill grouping with getActiveFormat-driven dropdowns
The highlighter and font-size features grouped their per-color/size formats
into one toolbar dropdown via a custom createSlotFill, which reinvented
Gutenberg's own RichText.ToolbarControls slot purely to aggregate each
format's active state across independently-rendered edits.

Read each item's active state directly with getActiveFormat instead, so a
single format's edit can render the whole ToolbarDropdownMenu without any
slot fill. The menu items are rendered as MenuGroup/MenuItem children rather
than via the `controls` prop, which keeps the styled-title preview typed as
ReactNode and removes the `as unknown as string` casts.

The shared helper is dropped and the logic is inlined into each feature.
