--- src/font-size.ts
@@ -1,19 +1,0 @@
-/**
- * WordPress dependencies
- */
-import { __ } from '@wordpress/i18n';
-import { textColor as icon } from '@wordpress/icons';
-
-/**
- * Internal dependencies
- */
-import { getRichTextSetting } from './helper';
-import { registerFormatType } from './register-format-type';
-
-const label = __( 'Font size', 'richtext-extension' );
-const slotFillName = 'FontSizeDropdownControls';
-
-rtexConf.fontSize.forEach( ( { title, className }, index ) => {
-	const setting = getRichTextSetting( { label, icon, title, className, slotFillName }, index );
-	registerFormatType( ...setting );
-} );
