+++ src/types.ts
@@ -9,14 +9,6 @@ export type FormatEditProps = {
 	onChange: ( value: RichTextValue ) => void;
 };
 
-export type RichTextSettingArgs = {
-	label: string;
-	icon: JSX.Element;
-	title: string;
-	className: string;
-	slotFillName: string;
-};
-
 export type FormatTypeSettings = {
 	title: string;
 	tagName: string;
